Add files via upload
Goal of this project -->  Creating a classifier to find that a message is spam or not.
First, text mining is used to extract meaningful information from the data set in order to make the information contained in the text accessible to the various algorithms. After that Lasso, Ridge, and ElasticNet algorithms were used. Final the accuracy of each model was compared in order to see that which method has better performance.